Add Kotlin API dump plugin
Unfortunately this plugin does not actually validate that changes are compatible so reviewers must still remain vigilant to changes to this file.

Double unfortunately, the plugin only validates the binary surface of Java bytecode-based targets. It does not support source/IR compatibility of common targets, nor targets other than the JVM at any source/IR/binary layer. Thankfully our
sources are 100% common and thus the JVM binary surface matches the common surface, but that's not guaranteed to stay.

Long term we deserve something better than this, but this is good enough for now.