Support CoroutineScope without a Job in testIn
Most `CoroutineScope`s contain a `Job`, and should contain a `Job` to work with structured concurrency. However, some `CoroutineScope`s do not have a `Job` (`GlobalScope` is an example).

This change adds support for `CoroutineScope`s with no `Job` to Turbine's `testIn` function, which previously would throw `IllegalStateException` if called with such a `CoroutineScope`.

A `CoroutineScope` with no `Job` never completes, so a call to `invokeOnCompletion` is not needed in that case.